requirements.txt
notion-client==2.2.1
requests==2.31.0
python-dotenv==1.0.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
tzdata
anthropic
